+++ tests/Admin/OptionsTest.php
@@ -3,8 +3,8 @@
 namespace ImageConverterWebP\Tests\Admin;
 
 use Mockery;
-use WP_Mock\Tools\TestCase;
 use ImageConverterWebP\Admin\Options;
+use Badasswp\WPMockTC\WPMockTestCase;
 
 /**
  * @covers \ImageConverterWebP\Admin\Options::get_form_page
@@ -12,26 +12,16 @@ use ImageConverterWebP\Admin\Options;
  * @covers \ImageConverterWebP\Admin\Options::get_form_notice
  * @covers \ImageConverterWebP\Admin\Options::get_form_fields
  */
-class OptionsTest extends TestCase {
+class OptionsTest extends WPMockTestCase {
 	public function setUp(): void {
-		\WP_Mock::setUp();
+		parent::setUp();
 	}
 
 	public function tearDown(): void {
-		\WP_Mock::tearDown();
+		parent::tearDown();
 	}
 
 	public function test_get_form_page() {
-		\WP_Mock::userFunction(
-			'esc_html__',
-			[
-				'times'  => 2,
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
 		$form_page = Options::get_form_page();
 
 		$this->assertSame(
@@ -46,16 +36,6 @@ class OptionsTest extends TestCase {
 	}
 
 	public function test_get_form_submit() {
-		\WP_Mock::userFunction(
-			'esc_html__',
-			[
-				'times'  => 2,
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
 		$form_submit = Options::get_form_submit();
 
 		$this->assertSame(
@@ -75,33 +55,6 @@ class OptionsTest extends TestCase {
 	}
 
 	public function test_get_form_fields() {
-		\WP_Mock::userFunction(
-			'esc_html__',
-			[
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
-		\WP_Mock::userFunction(
-			'esc_attr',
-			[
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
-		\WP_Mock::userFunction(
-			'esc_attr__',
-			[
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
 		$form_fields = Options::get_form_fields();
 
 		$this->assertSame(
@@ -160,16 +113,6 @@ class OptionsTest extends TestCase {
 	}
 
 	public function test_get_form_notice() {
-		\WP_Mock::userFunction(
-			'esc_html__',
-			[
-				'times'  => 1,
-				'return' => function ( $text, $domain = 'image-converter-webp' ) {
-					return $text;
-				},
-			]
-		);
-
 		$form_notice = Options::get_form_notice();
 
 		$this->assertSame(