+++ tests/Services/BootTest.php
@@ -2,7 +2,9 @@
 
 namespace ImageConverterWebP\Tests\Services;
 
+use WP_Mock;
 use Mockery;
+use ReflectionClass;
 use WP_Mock\Tools\TestCase;
 use ImageConverterWebP\Services\Boot;
 use ImageConverterWebP\Abstracts\Service;
@@ -17,17 +19,17 @@ class BootTest extends TestCase {
 	public Boot $boot;
 
 	public function setUp(): void {
-		\WP_Mock::setUp();
+		WP_Mock::setUp();
 
 		$this->boot = new Boot();
 	}
 
 	public function tearDown(): void {
-		\WP_Mock::tearDown();
+		WP_Mock::tearDown();
 	}
 
 	public function test_register() {
-		\WP_Mock::expectActionAdded( 'init', [ $this->boot, 'register_translation' ] );
+		WP_Mock::expectActionAdded( 'init', [ $this->boot, 'register_translation' ] );
 
 		$this->boot->register();
 
@@ -35,14 +37,14 @@ class BootTest extends TestCase {
 	}
 
 	public function test_register_translation() {
-		$boot = new \ReflectionClass( Boot::class );
+		$boot = new ReflectionClass( Boot::class );
 
-		\WP_Mock::userFunction( 'plugin_basename' )
+		WP_Mock::userFunction( 'plugin_basename' )
 			->once()
 			->with( $boot->getFileName() )
 			->andReturn( '/inc/Services/Boot.php' );
 
-		\WP_Mock::userFunction( 'load_plugin_textdomain' )
+		WP_Mock::userFunction( 'load_plugin_textdomain' )
 			->once()
 			->with(
 				'image-converter-webp',