+++ tests/Services/MainTest.php
@@ -2,6 +2,8 @@
 
 namespace ImageConverterWebP\Tests\Services;
 
+use WP_Mock;
+use WP_Post;
 use Mockery;
 use WP_Mock\Tools\TestCase;
 use ImageConverterWebP\Services\Main;
@@ -16,7 +18,6 @@ use ImageConverterWebP\Core\Converter;
  * @covers \ImageConverterWebP\Services\Main::register_webp_img_deletion
  * @covers \ImageConverterWebP\Services\Main::register_webp_attachment_fields
  * @covers \ImageConverterWebP\Services\Main::show_webp_images_on_wp_media_modal
- * @covers \ImageConverterWebP\Services\Main::get_webp_metadata
  * @covers icfw_get_settings
  * @covers icfw_get_equivalent
  * @covers icfw_get_abs_image
@@ -25,21 +26,50 @@ class MainTest extends TestCase {
 	public Main $main;
 
 	public function setUp(): void {
-		\WP_Mock::setUp();
+		WP_Mock::setUp();
 
 		$this->main = new Main();
 	}
 
 	public function tearDown(): void {
-		\WP_Mock::tearDown();
+		WP_Mock::tearDown();
 	}
 
 	public function test_register() {
-		\WP_Mock::expectActionAdded( 'add_attachment', [ $this->main, 'register_webp_img_creation' ], 10, 1 );
-		\WP_Mock::expectFilterAdded( 'wp_generate_attachment_metadata', [ $this->main, 'register_webp_img_srcset_creation' ], 10, 3 );
-		\WP_Mock::expectActionAdded( 'delete_attachment', [ $this->main, 'register_webp_img_deletion' ], 10, 1 );
-		\WP_Mock::expectFilterAdded( 'attachment_fields_to_edit', [ $this->main, 'register_webp_attachment_fields' ], 10, 2 );
-		\WP_Mock::expectFilterAdded( 'wp_prepare_attachment_for_js', [ $this->main, 'show_webp_images_on_wp_media_modal' ], 10, 3 );
+		WP_Mock::expectActionAdded(
+			'add_attachment',
+			[ $this->main, 'register_webp_img_creation' ],
+			10,
+			1
+		);
+
+		WP_Mock::expectFilterAdded(
+			'wp_generate_attachment_metadata',
+			[ $this->main, 'register_webp_img_srcset_creation' ],
+			10,
+			3
+		);
+
+		WP_Mock::expectActionAdded(
+			'delete_attachment',
+			[ $this->main, 'register_webp_img_deletion' ],
+			10,
+			1
+		);
+
+		WP_Mock::expectFilterAdded(
+			'attachment_fields_to_edit',
+			[ $this->main, 'register_webp_attachment_fields' ],
+			10,
+			2
+		);
+
+		WP_Mock::expectFilterAdded(
+			'wp_prepare_attachment_for_js',
+			[ $this->main, 'show_webp_images_on_wp_media_modal' ],
+			10,
+			3
+		);
 
 		$this->main->register();
 
@@ -50,7 +80,7 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_get_attachment_url' )
+		WP_Mock::userFunction( 'wp_get_attachment_url' )
 			->once()
 			->with( 1 )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg' );
@@ -60,7 +90,7 @@ class MainTest extends TestCase {
 			'url' => 'https://example.com/wp-content/uploads/2024/01/sample.jpeg',
 		];
 
-		\WP_Mock::userFunction( 'get_option' )
+		WP_Mock::userFunction( 'get_option' )
 			->once()
 			->with( 'icfw', [] )
 			->andReturn(
@@ -87,7 +117,7 @@ class MainTest extends TestCase {
 			'url' => 'https://example.com/wp-content/uploads/2024/01/sample.jpeg',
 		];
 
-		\WP_Mock::userFunction( 'wp_get_attachment_url' )
+		WP_Mock::userFunction( 'wp_get_attachment_url' )
 			->once()
 			->with( 1 )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg' );
@@ -96,7 +126,7 @@ class MainTest extends TestCase {
 			->once()
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.webp' );
 
-		\WP_Mock::userFunction( 'get_option' )
+		WP_Mock::userFunction( 'get_option' )
 			->once()
 			->with( 'icfw', [] )
 			->andReturn(
@@ -131,12 +161,12 @@ class MainTest extends TestCase {
 			],
 		];
 
-		\WP_Mock::userFunction( 'wp_get_attachment_image_url' )
+		WP_Mock::userFunction( 'wp_get_attachment_image_url' )
 			->once()
 			->with( 1 )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg' );
 
-		\WP_Mock::userFunction( 'trailingslashit' )
+		WP_Mock::userFunction( 'trailingslashit' )
 			->times( 3 )
 			->with( 'https://example.com/wp-content/uploads/2024/01' )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/' );
@@ -144,7 +174,7 @@ class MainTest extends TestCase {
 		$main->converter->shouldReceive( 'convert' )
 			->times( 3 );
 
-		\WP_Mock::userFunction( 'get_option' )
+		WP_Mock::userFunction( 'get_option' )
 			->times( 3 )
 			->with( 'icfw', [] )
 			->andReturn(
@@ -176,17 +206,17 @@ class MainTest extends TestCase {
 			],
 		];
 
-		\WP_Mock::userFunction( 'wp_get_attachment_image_url' )
+		WP_Mock::userFunction( 'wp_get_attachment_image_url' )
 			->once()
 			->with( 1 )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg' );
 
-		\WP_Mock::userFunction( 'trailingslashit' )
+		WP_Mock::userFunction( 'trailingslashit' )
 			->times( 3 )
 			->with( 'https://example.com/wp-content/uploads/2024/01' )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/' );
 
-		\WP_Mock::userFunction( 'get_option' )
+		WP_Mock::userFunction( 'get_option' )
 			->times( 3 )
 			->with( 'icfw', [] )
 			->andReturn(
@@ -204,7 +234,7 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( false );
@@ -218,17 +248,17 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->once()
 			->with( 1 )
 			->andReturn( '' );
 
-		\WP_Mock::userFunction( 'wp_get_attachment_metadata' )
+		WP_Mock::userFunction( 'wp_get_attachment_metadata' )
 			->once()
 			->with( 1 )
 			->andReturn( [] );
@@ -242,19 +272,19 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->once()
 			->with( 1 )
 			->andReturn( __DIR__ . '/sample.jpeg' );
 
-		\WP_Mock::expectAction( 'icfw_delete', __DIR__ . '/sample.webp', 1 );
+		WP_Mock::expectAction( 'icfw_delete', __DIR__ . '/sample.webp', 1 );
 
-		\WP_Mock::userFunction( 'wp_get_attachment_metadata' )
+		WP_Mock::userFunction( 'wp_get_attachment_metadata' )
 			->once()
 			->with( 1 )
 			->andReturn( [] );
@@ -271,17 +301,17 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->once()
 			->with( 1 )
 			->andReturn( '/sample.jpeg' );
 
-		\WP_Mock::userFunction( 'wp_get_attachment_metadata' )
+		WP_Mock::userFunction( 'wp_get_attachment_metadata' )
 			->once()
 			->with( 1 )
 			->andReturn( [] );
@@ -295,19 +325,19 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->once()
 			->with( 1 )
 			->andReturn( __DIR__ . '/sample.jpeg' );
 
-		\WP_Mock::expectAction( 'icfw_delete', __DIR__ . '/sample.webp', 1 );
+		WP_Mock::expectAction( 'icfw_delete', __DIR__ . '/sample.webp', 1 );
 
-		\WP_Mock::userFunction(
+		WP_Mock::userFunction(
 			'trailingslashit',
 			[
 				'times'  => 3,
@@ -317,7 +347,7 @@ class MainTest extends TestCase {
 			]
 		);
 
-		\WP_Mock::userFunction( 'wp_get_attachment_metadata' )
+		WP_Mock::userFunction( 'wp_get_attachment_metadata' )
 			->once()
 			->with( 1 )
 			->andReturn(
@@ -336,9 +366,9 @@ class MainTest extends TestCase {
 				]
 			);
 
-		\WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample1.webp', 1 );
-		\WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample2.webp', 1 );
-		\WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample3.webp', 1 );
+		WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample1.webp', 1 );
+		WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample2.webp', 1 );
+		WP_Mock::expectAction( 'icfw_metadata_delete', __DIR__ . '/sample3.webp', 1 );
 
 		// Create Mock Images.
 		$this->create_mock_image( __DIR__ . '/sample.webp' );
@@ -355,17 +385,17 @@ class MainTest extends TestCase {
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->once()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->once()
 			->with( 1 )
 			->andReturn( '/sample.jpeg' );
 
-		\WP_Mock::userFunction(
+		WP_Mock::userFunction(
 			'trailingslashit',
 			[
 				'times'  => 3,
@@ -375,7 +405,7 @@ class MainTest extends TestCase {
 			]
 		);
 
-		\WP_Mock::userFunction( 'wp_get_attachment_metadata' )
+		WP_Mock::userFunction( 'wp_get_attachment_metadata' )
 			->once()
 			->with( 1 )
 			->andReturn(
@@ -400,10 +430,10 @@ class MainTest extends TestCase {
 	}
 
 	public function test_register_webp_attachment_fields_escapes_array_return_type() {
-		$post     = Mockery::mock( \WP_Post::class )->makePartial();
+		$post     = Mockery::mock( WP_Post::class )->makePartial();
 		$post->ID = 1;
 
-		\WP_Mock::userFunction(
+		WP_Mock::userFunction(
 			'__',
 			[
 				'return' => function ( $text, $domain = 'image-converter-webp' ) {
@@ -412,7 +442,7 @@ class MainTest extends TestCase {
 			]
 		);
 
-		\WP_Mock::userFunction( 'get_post_meta' )
+		WP_Mock::userFunction( 'get_post_meta' )
 			->once()
 			->with( 1, 'icfw_img', true )
 			->andReturn( [] );
@@ -436,10 +466,10 @@ class MainTest extends TestCase {
 	public function test_register_webp_attachment_fields() {
 		$webp = 'https://example.com/wp-content/uploads/2024/01/sample.webp';
 
-		$post     = Mockery::mock( \WP_Post::class )->makePartial();
+		$post     = Mockery::mock( WP_Post::class )->makePartial();
 		$post->ID = 1;
 
-		\WP_Mock::userFunction(
+		WP_Mock::userFunction(
 			'__',
 			[
 				'return' => function ( $text, $domain = 'image-converter-webp' ) {
@@ -448,7 +478,7 @@ class MainTest extends TestCase {
 			]
 		);
 
-		\WP_Mock::userFunction( 'get_post_meta' )
+		WP_Mock::userFunction( 'get_post_meta' )
 			->once()
 			->with( 1, 'icfw_img', true )
 			->andReturn( $webp );
@@ -470,14 +500,14 @@ class MainTest extends TestCase {
 	}
 
 	public function test_show_webp_images_on_wp_media_modal_bails_out_if_not_image() {
-		$attachment     = Mockery::mock( \WP_Post::class )->makePartial();
+		$attachment     = Mockery::mock( WP_Post::class )->makePartial();
 		$attachment->ID = 1;
 
-		\WP_Mock::userFunction( 'get_post_meta' )
+		WP_Mock::userFunction( 'get_post_meta' )
 			->with( 1, 'icfw_img', true )
 			->andReturn( 'https://example.com/wp-content/uploads/sample.pdf' );
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->with( 1 )
 			->andReturnUsing(
 				function ( $arg ) {
@@ -514,19 +544,19 @@ class MainTest extends TestCase {
 	}
 
 	public function test_show_webp_images_on_wp_media_modal_bails_out_if_webp_image_does_not_exist() {
-		$attachment     = Mockery::mock( \WP_Post::class )->makePartial();
+		$attachment     = Mockery::mock( WP_Post::class )->makePartial();
 		$attachment->ID = 1;
 
-		\WP_Mock::userFunction( 'get_post_meta' )
+		WP_Mock::userFunction( 'get_post_meta' )
 			->with( 1, 'icfw_img', true )
 			->andReturn( 'https://example.com/wp-content/uploads/image.webp' );
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->twice()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->with( 1 )
 			->andReturn( '/var/www/html/wp-content/uploads/image.jpeg' );
 
@@ -553,54 +583,24 @@ class MainTest extends TestCase {
 	}
 
 	public function test_show_webp_images_on_wp_media_modal_passes() {
-		$attachment     = Mockery::mock( \WP_Post::class )->makePartial();
+		$attachment     = Mockery::mock( WP_Post::class )->makePartial();
 		$attachment->ID = 1;
 
 		$main = Mockery::mock( Main::class )->makePartial();
 		$main->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'get_post_meta' )
+		WP_Mock::userFunction( 'get_post_meta' )
 			->with( 1, 'icfw_img', true )
-			->andReturn( __DIR__ . '/sample.webp' );
+			->andReturn( 'https://example.com/wp-content/uploads/image.webp' );
 
-		\WP_Mock::userFunction( 'wp_attachment_is_image' )
+		WP_Mock::userFunction( 'wp_attachment_is_image' )
 			->twice()
 			->with( 1 )
 			->andReturn( true );
 
-		\WP_Mock::userFunction( 'get_attached_file' )
+		WP_Mock::userFunction( 'get_attached_file' )
 			->with( 1 )
-			->andReturn( __DIR__ . '/sample.webp' );
-
-		$main->shouldReceive( 'get_webp_metadata' )
-			->with(
-				[
-					'sizes' => [
-						'thumbnail' => [
-							'url' => 'https://example.com/wp-content/uploads/image-150x150.jpeg',
-						],
-						'medium'    => [
-							'url' => 'https://example.com/wp-content/uploads/image-300x300.jpeg',
-						],
-						'large'     => [
-							'url' => 'https://example.com/wp-content/uploads/image-1024x1024.jpeg',
-						],
-						'full'      => [
-							'url' => __DIR__ . '/sample.webp',
-						],
-					],
-				]
-			)->andReturnUsing(
-				function ( $arg ) {
-					$types = [ 'thumbnail', 'medium', 'large' ];
-
-					foreach ( $types as $type ) {
-						$arg['sizes'][ $type ]['url'] = str_replace( '.jpeg', '.webp', $arg['sizes'][ $type ]['url'] );
-					}
-
-					return $arg;
-				}
-			);
+			->andReturn( __DIR__ . '/image.webp' );
 
 		$metadata = [
 			'sizes' => [
@@ -619,8 +619,6 @@ class MainTest extends TestCase {
 			],
 		];
 
-		$this->create_mock_image( __DIR__ . '/sample.webp' );
-
 		$metadata = $main->show_webp_images_on_wp_media_modal( $metadata, $attachment, false );
 
 		$this->assertSame(
@@ -631,55 +629,10 @@ class MainTest extends TestCase {
 						'url' => 'https://example.com/wp-content/uploads/image-150x150.webp',
 					],
 					'medium'    => [
-						'url' => 'https://example.com/wp-content/uploads/image-300x300.webp',
+						'url' => 'https://example.com/wp-content/uploads/image-300x300.jpeg',
 					],
 					'large'     => [
-						'url' => 'https://example.com/wp-content/uploads/image-1024x1024.webp',
-					],
-					'full'      => [
-						'url' => __DIR__ . '/sample.webp',
-					],
-				],
-			]
-		);
-		$this->assertConditionsMet();
-	}
-
-	public function test_get_webp_metadata() {
-		$main = Mockery::mock( Main::class )->makePartial();
-		$main->shouldAllowMockingProtectedMethods();
-
-		$metadata = [
-			'sizes' => [
-				'thumbnail' => [
-					'url' => 'https://example.com/wp-content/uploads/image-150x150.jpeg',
-				],
-				'medium'    => [
-					'url' => 'https://example.com/wp-content/uploads/image-300x300.jpeg',
-				],
-				'large'     => [
-					'url' => 'https://example.com/wp-content/uploads/image-1024x1024.jpeg',
-				],
-				'full'      => [
-					'url' => 'https://example.com/wp-content/uploads/image.webp',
-				],
-			],
-		];
-
-		$metadata = $main->get_webp_metadata( $metadata );
-
-		$this->assertSame(
-			$metadata,
-			[
-				'sizes' => [
-					'thumbnail' => [
-						'url' => 'https://example.com/wp-content/uploads/image-150x150.webp',
-					],
-					'medium'    => [
-						'url' => 'https://example.com/wp-content/uploads/image-300x300.webp',
-					],
-					'large'     => [
-						'url' => 'https://example.com/wp-content/uploads/image-1024x1024.webp',
+						'url' => 'https://example.com/wp-content/uploads/image-1024x1024.jpeg',
 					],
 					'full'      => [
 						'url' => 'https://example.com/wp-content/uploads/image.webp',