+++ tests/Services/PageLoadTest.php
@@ -2,8 +2,10 @@
 
 namespace ImageConverterWebP\Tests\Services;
 
+use WP_Mock;
 use Mockery;
-use WP_Mock\Tools\TestCase;
+use WP_Error;
+use Badasswp\WPMockTC\WPMockTestCase;
 use ImageConverterWebP\Core\Converter;
 use ImageConverterWebP\Services\PageLoad;
 
@@ -15,27 +17,45 @@ use ImageConverterWebP\Services\PageLoad;
  * @covers \ImageConverterWebP\Services\PageLoad::register_wp_get_attachment_image
  * @covers \ImageConverterWebP\Services\PageLoad::register_post_thumbnail_html
  * @covers \ImageConverterWebP\Services\PageLoad::get_webp_image_html
- * @covers \ImageConverterWebP\Services\PageLoad::_get_webp_html
+ * @covers \ImageConverterWebP\Services\PageLoad::get_webp
+ * @covers \ImageConverterWebP\Services\PageLoad::get_all_srcset_images
  * @covers icfw_get_settings
  */
-class PageLoadTest extends TestCase {
+class PageLoadTest extends WPMockTestCase {
 	public array $source;
 	public PageLoad $page_load;
 
 	public function setUp(): void {
-		\WP_Mock::setUp();
+		parent::setUp();
 
 		$this->page_load = new PageLoad();
 	}
 
 	public function tearDown(): void {
-		\WP_Mock::tearDown();
+		parent::tearDown();
 	}
 
 	public function test_register() {
-		\WP_Mock::expectFilterAdded( 'render_block', [ $this->page_load, 'register_render_block' ], 20, 2 );
-		\WP_Mock::expectFilterAdded( 'wp_get_attachment_image', [ $this->page_load, 'register_wp_get_attachment_image' ], 10, 5 );
-		\WP_Mock::expectFilterAdded( 'post_thumbnail_html', [ $this->page_load, 'register_post_thumbnail_html' ], 10, 5 );
+		WP_Mock::expectFilterAdded(
+			'render_block',
+			[ $this->page_load, 'register_render_block' ],
+			20,
+			2
+		);
+
+		WP_Mock::expectFilterAdded(
+			'wp_get_attachment_image',
+			[ $this->page_load, 'register_wp_get_attachment_image' ],
+			10,
+			5
+		);
+
+		WP_Mock::expectFilterAdded(
+			'post_thumbnail_html',
+			[ $this->page_load, 'register_post_thumbnail_html' ],
+			10,
+			5
+		);
 
 		$this->page_load->register();
 
@@ -93,7 +113,7 @@ class PageLoadTest extends TestCase {
 			->with( '<img src="sample.jpeg"/>', 1 )
 			->andReturn( '<img src="sample.webp"/>' );
 
-		\WP_Mock::onFilter( 'icfw_attachment_html' )
+		WP_Mock::onFilter( 'icfw_attachment_html' )
 			->with(
 				'<img src="sample.webp"/>',
 				1
@@ -124,7 +144,7 @@ class PageLoadTest extends TestCase {
 			->with( '<img src="sample.jpeg"/>', 2 )
 			->andReturn( '<img src="sample.webp"/>' );
 
-		\WP_Mock::onFilter( 'icfw_thumbnail_html' )
+		WP_Mock::onFilter( 'icfw_thumbnail_html' )
 			->with(
 				'<img src="sample.webp"/>',
 				2
@@ -139,7 +159,7 @@ class PageLoadTest extends TestCase {
 		$this->assertConditionsMet();
 	}
 
-	public function test_get_webp_image_html_returns_emtpy_image_if_empty() {
+	public function test_get_webp_image_html_returns_default_html_if_empty() {
 		$page_load = Mockery::mock( PageLoad::class )->makePartial();
 		$page_load->shouldAllowMockingProtectedMethods();
 
@@ -149,7 +169,7 @@ class PageLoadTest extends TestCase {
 		$this->assertConditionsMet();
 	}
 
-	public function test_get_webp_image_html_returns_html_if_no_image_in_html() {
+	public function test_get_webp_image_html_returns_default_html_if_no_image_in_html() {
 		$page_load = Mockery::mock( PageLoad::class )->makePartial();
 		$page_load->shouldAllowMockingProtectedMethods();
 
@@ -159,45 +179,38 @@ class PageLoadTest extends TestCase {
 		$this->assertConditionsMet();
 	}
 
-	public function test_get_webp_image_html_returns_original_html_if_no_image_src_is_in_html() {
+	public function test_get_webp_image_html_returns_default_html_if_page_load_is_not_activated() {
 		$page_load = Mockery::mock( PageLoad::class )->makePartial();
 		$page_load->shouldAllowMockingProtectedMethods();
 
-		$image = $page_load->get_webp_image_html( '<figure><img/></figure>' );
-
-		$this->assertSame( '<figure><img/></figure>', $image );
-		$this->assertConditionsMet();
-	}
-
-	public function test_get_webp_image_html_returns_original_html_if_image_is_webp() {
-		$page_load = Mockery::mock( PageLoad::class )->makePartial();
-		$page_load->shouldAllowMockingProtectedMethods();
-
-		$image = $page_load->get_webp_image_html( '<figure><img src="john.webp"/></figure>' );
-
-		$this->assertSame( '<figure><img src="john.webp"/></figure>', $image );
-		$this->assertConditionsMet();
-	}
-
-	public function test_get_webp_html_bails_out_and_returns_same_image_html() {
-		$page_load = Mockery::mock( PageLoad::class )->makePartial();
-		$page_load->shouldAllowMockingProtectedMethods();
-
-		$page_load->converter = Mockery::mock( Converter::class )->makePartial();
-		$page_load->converter->shouldAllowMockingProtectedMethods();
-
-		$error = Mockery::mock( \WP_Error::class )->makePartial();
-
-		$page_load->converter->shouldReceive( 'convert' )
-			->once()->with()
-			->andReturn( $error );
-
-		\WP_Mock::userFunction( 'is_wp_error' )
+		WP_Mock::userFunction( 'get_option' )
 			->once()
-			->with( $error )
-			->andReturn( true );
+			->with( 'icfw', [] )
+			->andReturn(
+				[
+					'page_load' => false,
+				]
+			);
 
-		\WP_Mock::userFunction( 'get_option' )
+		$image = $page_load->get_webp_image_html( '<figure><img src="john.png"/></figure>' );
+
+		$this->assertSame( '<figure><img src="john.png"/></figure>', $image );
+		$this->assertConditionsMet();
+	}
+
+	public function test_get_webp_image_html_works_correctly_and_returns_new_html_with_webp_images() {
+		$page_load = Mockery::mock( PageLoad::class )->makePartial();
+		$page_load->shouldAllowMockingProtectedMethods();
+
+		$page_load->shouldReceive( 'get_webp' )
+			->andReturnUsing(
+				function ( $arg1, $arg2 ) {
+					$ext = pathinfo( $arg1, PATHINFO_EXTENSION );
+					return str_replace( $ext, 'webp', $arg1 );
+				}
+			);
+
+		WP_Mock::userFunction( 'get_option' )
 			->once()
 			->with( 'icfw', [] )
 			->andReturn(
@@ -206,50 +219,96 @@ class PageLoadTest extends TestCase {
 				]
 			);
 
-		$img_html = $page_load->_get_webp_html( 'https://example.com/wp-content/uploads/2024/01/sample.pdf', '<img src="https://example.com/wp-content/uploads/2024/01/sample.pdf"/>', 1 );
+		WP_Mock::userFunction( 'wp_filter_content_tags' )
+			->with( '<figure class="wp-block-image size-large"><img src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140"/></figure>' )
+			->andReturn( '<figure class="wp-block-image size-large"><img width="799" height="1024" src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140" srcset="https://example.com/wp-content/uploads/2025/12/image-799x1024.png 799w, https://example.com/wp-content/uploads/2025/12/image-234x300.png 234w, https://example.com/wp-content/uploads/2025/12/image-768x985.png 768w, https://example.com/wp-content/uploads/2025/12/image.png 922w" sizes="(max-width: 799px) 100vw, 799px" /></figure>' );
 
-		$this->assertSame( $img_html, '<img src="https://example.com/wp-content/uploads/2024/01/sample.pdf"/>' );
+		$image = $page_load->get_webp_image_html( '<figure class="wp-block-image size-large"><img src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140"/></figure>' );
+
+		$this->assertSame(
+			$image,
+			'<figure class="wp-block-image size-large"><img width="799" height="1024" src="https://example.com/wp-content/uploads/2025/12/image-799x1024.webp" alt="" class="wp-image-1140" srcset="https://example.com/wp-content/uploads/2025/12/image-799x1024.webp 799w, https://example.com/wp-content/uploads/2025/12/image-234x300.webp 234w, https://example.com/wp-content/uploads/2025/12/image-768x985.webp 768w, https://example.com/wp-content/uploads/2025/12/image.webp 922w" sizes="(max-width: 799px) 100vw, 799px" /></figure>'
+		);
 		$this->assertConditionsMet();
 	}
 
-	public function test_get_webp_html_returns_new_image_html() {
+	public function test_get_webp_image_html_fails_gracefully_and_returns_default_html_with_original_images() {
 		$page_load = Mockery::mock( PageLoad::class )->makePartial();
 		$page_load->shouldAllowMockingProtectedMethods();
 
 		$page_load->converter = Mockery::mock( Converter::class )->makePartial();
 		$page_load->converter->shouldAllowMockingProtectedMethods();
 
-		$this->create_mock_image( __DIR__ . '/sample.webp' );
-		$page_load->converter->abs_dest = __DIR__ . '/sample.webp';
+		$page_load->converter->shouldReceive( 'convert' )
+			->andReturn( Mockery::mock( WP_Error::class )->makePartial() );
 
-		$error = Mockery::mock( \WP_Error::class )->makePartial();
+		WP_Mock::userFunction( 'get_option' )
+			->once()
+			->with( 'icfw', [] )
+			->andReturn(
+				[
+					'page_load' => true,
+				]
+			);
 
-		$this->source['url'] = 'https://example.com/wp-content/uploads/2024/01/sample.jpeg';
+		WP_Mock::userFunction( 'wp_filter_content_tags' )
+			->with( '<figure class="wp-block-image size-large"><img src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140"/></figure>' )
+			->andReturn( '<figure class="wp-block-image size-large"><img width="799" height="1024" src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140" srcset="https://example.com/wp-content/uploads/2025/12/image-799x1024.png 799w, https://example.com/wp-content/uploads/2025/12/image-234x300.png 234w, https://example.com/wp-content/uploads/2025/12/image-768x985.png 768w, https://example.com/wp-content/uploads/2025/12/image.png 922w" sizes="(max-width: 799px) 100vw, 799px" /></figure>' );
+
+		$image = $page_load->get_webp_image_html( '<figure class="wp-block-image size-large"><img src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140"/></figure>', 1 );
+
+		$this->assertSame(
+			$image,
+			'<figure class="wp-block-image size-large"><img width="799" height="1024" src="https://example.com/wp-content/uploads/2025/12/image-799x1024.png" alt="" class="wp-image-1140" srcset="https://example.com/wp-content/uploads/2025/12/image-799x1024.png 799w, https://example.com/wp-content/uploads/2025/12/image-234x300.png 234w, https://example.com/wp-content/uploads/2025/12/image-768x985.png 768w, https://example.com/wp-content/uploads/2025/12/image.png 922w" sizes="(max-width: 799px) 100vw, 799px" /></figure>'
+		);
+		$this->assertConditionsMet();
+	}
+
+	public function test_get_webp_bails_out_if_it_returns_wp_error_and_returns_default_image() {
+		$page_load = Mockery::mock( PageLoad::class )->makePartial();
+		$page_load->shouldAllowMockingProtectedMethods();
+
+		$page_load->converter = Mockery::mock( Converter::class )->makePartial();
+		$page_load->converter->shouldAllowMockingProtectedMethods();
 
 		$page_load->converter->shouldReceive( 'convert' )
-			->once()->with()
+			->andReturn( Mockery::mock( WP_Error::class )->makePartial() );
+
+		$img_html = $page_load->get_webp( 'https://example.com/wp-content/uploads/2024/01/sample.png', 1 );
+
+		$this->assertSame( $img_html, 'https://example.com/wp-content/uploads/2024/01/sample.png' );
+		$this->assertConditionsMet();
+	}
+
+	public function test_get_webp_returns_new_image_html() {
+		$page_load = Mockery::mock( PageLoad::class )->makePartial();
+		$page_load->shouldAllowMockingProtectedMethods();
+
+		$page_load->converter = Mockery::mock( Converter::class )->makePartial();
+		$page_load->converter->shouldAllowMockingProtectedMethods();
+
+		$page_load->converter->shouldReceive( 'convert' )
 			->andReturn( 'https://example.com/wp-content/uploads/2024/01/sample.webp' );
 
-		\WP_Mock::userFunction( 'is_wp_error' )
-			->once()
-			->with( 'https://example.com/wp-content/uploads/2024/01/sample.webp' )
-			->andReturn( false );
+		$img_html = $page_load->get_webp( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg', 1 );
 
-		\WP_Mock::userFunction( 'get_option' )
-			->once()
-			->with( 'icfw', [] )
-			->andReturn(
-				[
-					'page_load' => true,
-				]
-			);
-
-		$img_html = $page_load->_get_webp_html( 'https://example.com/wp-content/uploads/2024/01/sample.jpeg', '<img src="https://example.com/wp-content/uploads/2024/01/sample.jpeg"/>', 1 );
-
-		$this->assertSame( $img_html, '<img src="https://example.com/wp-content/uploads/2024/01/sample.webp"/>' );
+		$this->assertSame( $img_html, 'https://example.com/wp-content/uploads/2024/01/sample.webp' );
 		$this->assertConditionsMet();
+	}
 
-		$this->destroy_mock_image( __DIR__ . '/sample.webp' );
+	public function test_get_all_srcset_images() {
+		$page_load = Mockery::mock( PageLoad::class )->makePartial();
+		$page_load->shouldAllowMockingProtectedMethods();
+
+		$this->assertSame(
+			[
+				'https://example.com/wp-content/uploads/2025/12/image.webp',
+				'https://example.com/wp-content/uploads/2025/12/image-234x300.png',
+				'https://example.com/wp-content/uploads/2025/12/image-799x1024.png',
+				'https://example.com/wp-content/uploads/2025/12/image-768x985.png',
+			],
+			$page_load->get_all_srcset_images( 'https://example.com/wp-content/uploads/2025/12/image.webp 922w, https://example.com/wp-content/uploads/2025/12/image-234x300.png 234w, https://example.com/wp-content/uploads/2025/12/image-799x1024.png 799w, https://example.com/wp-content/uploads/2025/12/image-768x985.png 768w' )
+		);
 	}
 
 	public function create_mock_image( $image_file_name ) {